+++ packages/ui/src/components/property-editor/editor-items/DividerItem.jsx
@@ -1,0 +1,6 @@
+import { Divider } from "../../divider";
+
+export function DividerItem({ item }) {
+  const { margin = 10, style } = item;
+  return <Divider style={style} margin={margin} />;
+}